fix: PHPStan level 7 — type safety (96 → 54)
- Add fopen() false checks in ExportController, TaxExport, ProductCsvService,
  BackupDatabases, SurveyResults
- Cast QrCode color args and imagecolorallocate args to int
- Cast float|int|string to float for property assignments (FinanceSummary)
- Fix ucfirst on enum types (use ->value)
- Cast mktime return to int, filesize to int
- Handle glob() returning false (?: [])
- Cast ob_get_clean() and hash_hmac payload to string
- Fix number_format on mixed numeric types
- Bump PHPStan to level 7

+++ app/Services/ReportService.php
@@ -165,7 +165,7 @@ class ReportService
             $mInc = (float) ($monthlyIncome[$m] ?? 0);
             $mExp = (float) ($monthlyExpenses[$m] ?? 0);
             $monthly->push([
-                'month' => date('M', mktime(0, 0, 0, $m, 1)),
+                'month' => date('M', (int) mktime(0, 0, 0, $m, 1)),
                 'revenue' => $mRev + $mInc,
                 'expenses' => $mExp,
                 'profit' => $mRev + $mInc - $mExp,
